requirements.txt
openai==1.43.0
pinecone-client==5.0.1
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.23
langchain-experimental==0.0.65
langchainhub==0.1.21
wikipedia==1.4.0
duckduckgo-search==6.2.10
tiktoken==0.7.0
google-generativeai==0.5.4
langchain-google-genai==1.0.6
streamlit==1.38.0
streamlit-chat==0.1.1
pypdf==4.3.1
docx2txt==0.8
chromadb==0.5.3
langchain-chroma==0.1.3
python-dotenv